requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
PyMuPDF==1.24.0
Pillow==10.1.0
docx2pdf==0.1.8
python-docx==1.2.0
openai==1.3.0
pydantic==2.5.0
python-multipart==0.0.6
streamlit>=1.24.0
requests>=2.31.0
